partially fixed execution bug

+++ engine.py
@@ -24,20 +24,21 @@ def add_order_orderbook(new_order_id):
 
   orderbook_for_price = db.execute("SELECT * FROM orderbook WHERE price = :price", price=order["price"])
 
-  try_execution(order)
+  if try_execution(order) == True:
+    print("match1")
+  else:
+    # if there are no order at that price
+    if len(orderbook_for_price) == 0:
+      db.execute("INSERT INTO orderbook (pair,price,quantity,type) VALUES (?,?,?,?)", order["pair"], order["price"], order["quantity"], order["type"])
+      
+    elif len(orderbook_for_price) == 1:
 
-  # if there are no order at that price
-  if len(orderbook_for_price) == 0:
-    db.execute("INSERT INTO orderbook (pair,price,quantity,type) VALUES (?,?,?,?)", order["pair"], order["price"], order["quantity"], order["type"])
-  
-  elif len(orderbook_for_price) == 1:
-
-    if orderbook_for_price[0]["type"] == order["type"]:
-      db.execute("UPDATE orderbook SET quantity = :quantity WHERE pair = :pair AND price = :price AND type = :type", quantity = order["quantity"]+orderbook_for_price[0]["quantity"], pair = order["pair"], price = order["price"], type = order["type"] )
-    else:
-      print(order["order_id"] + " MATCHED")
-      db.execute("DELETE FROM open_orders WHERE user_id = :user AND order_id = :order_id", user=session["user_id"], order_id = new_order_id)
-      # if the price is the same and type is different then it means that someone is buying/selling for our desired price
+      if orderbook_for_price[0]["type"] == order["type"]:
+        db.execute("UPDATE orderbook SET quantity = :quantity WHERE pair = :pair AND price = :price AND type = :type", quantity = order["quantity"]+orderbook_for_price[0]["quantity"], pair = order["pair"], price = order["price"], type = order["type"] )
+      else:
+        print(order["order_id"] + " MATCHED")
+        db.execute("DELETE FROM open_orders WHERE user_id = :user AND order_id = :order_id", user=session["user_id"], order_id = new_order_id)
+        # if the price is the same and type is different then it means that someone is buying/selling for our desired price
 
 
 
@@ -58,11 +59,15 @@ def try_execution(order):
         if buy_order["quantity_left"] >= order_quantity_left:
           print(f"was enough {order}")
           db.execute("UPDATE hidden_orderbook SET quantity_left = :quantity WHERE order_id = :order_id", quantity = buy_order["quantity_left"]-order_quantity_left, order_id = buy_order["order_id"])
+          db.execute("DELETE FROM open_orders WHERE order_id = :orderid" , orderid= order["order_id"])
+          db.execute("UPDATE open_orders SET filled = :filled WHERE order_id = :order_id", filled = buy_order["quantity_left"]-order_quantity_left, order_id = buy_order["order_id"])
           order_quantity_left = 0
           orderbook_sync()
-          break
+          return True
         else:
           print("Wasn't enough")
+          return False
+
 
 
 
@@ -165,7 +170,7 @@ def orderbook_sync():
         db.execute("UPDATE orderbook SET quantity= :quantity WHERE price=:price AND type = :type AND pair = :pair", quantity = query["quantity"]+quantity_left, price = order["price"], type = order["type"], pair= order["pair"])
 
       else:
-        db.execute("INSERT INTO orderbook (pair,price,quantity,type) VALUES(?,?,?,?)", order["pair"], order["price"], order["quantity"], order["type"])
+        db.execute("INSERT INTO orderbook (pair,price,quantity,type) VALUES(?,?,?,?)", order["pair"], order["price"], order["quantity"]-order["filled"], order["type"])
         list_prices_already_on_orderbook.append(order["price"])
         print(str(order["price"]) + " added to orderbook")
